Separate spi from jshardware.c - still arduino lib
